+++ payments-ui-core/src/test/java/com/stripe/android/ui/core/forms/resources/LpmRepositoryTest.kt
@@ -3,6 +3,7 @@ package com.stripe.android.ui.core.forms.resources
 import android.app.Application
 import androidx.test.core.app.ApplicationProvider
 import com.google.common.truth.Truth.assertThat
+import com.stripe.android.payments.financialconnections.IsFinancialConnectionsAvailable
 import com.stripe.android.paymentsheet.forms.Delayed
 import com.stripe.android.ui.core.R
 import com.stripe.android.ui.core.elements.EmptyFormSpec
@@ -14,7 +15,12 @@ import java.util.Locale
 @RunWith(RobolectricTestRunner::class)
 class LpmRepositoryTest {
     private val lpmRepository = LpmRepository(
-        ApplicationProvider.getApplicationContext<Application>().resources
+        ApplicationProvider.getApplicationContext<Application>().resources,
+        object : IsFinancialConnectionsAvailable {
+            override fun invoke(): Boolean {
+                return true
+            }
+        }
     )
 
     @Test
@@ -133,5 +139,44 @@ class LpmRepositoryTest {
         assertThat(
             lpmRepository.fromCode("sofort")?.requirement?.piRequirements
         ).contains(Delayed)
+    }
+
+    @Test
+    fun `Verify that us_bank_account is supported when financial connections sdk available`() {
+        lpmRepository.initialize(
+            """
+              [
+                {
+                  "type": "us_bank_account"
+                }
+              ]
+            """.trimIndent().byteInputStream()
+        )
+
+        assertThat(lpmRepository.fromCode("us_bank_account")).isNotNull()
+    }
+
+    @Test
+    fun `Verify that us_bank_account not supported when financial connections sdk not available`() {
+        val lpmRepository = LpmRepository(
+            ApplicationProvider.getApplicationContext<Application>().resources,
+            object : IsFinancialConnectionsAvailable {
+                override fun invoke(): Boolean {
+                    return false
+                }
+            }
+        )
+
+        lpmRepository.initialize(
+            """
+              [
+                {
+                  "type": "us_bank_account"
+                }
+              ]
+            """.trimIndent().byteInputStream()
+        )
+
+        assertThat(lpmRepository.fromCode("us_bank_account")).isNull()
     }
 }